uploading the current debug notebook to work on another machine
